devices/sound: moved CEM devices to dedicated files. * CEM3310, CEM3320, CEM3360
