+++ src/devices/sound/va_vca.cpp
@@ -209,20 +209,6 @@ float ca3280_vca_lin_device::i_d(float r, float v_r, float v_minus)
 }
 
 
-cem3360_vca_device::cem3360_vca_device(const machine_config &mconfig, const char *tag, device_t *owner, uint32_t clock)
-	: va_vca_device(mconfig, CEM3360_VCA, tag, owner, clock)
-{
-}
-
-float cem3360_vca_device::cv_to_gain(float cv) const
-{
-	// Typical linear CV for max gain, as reported on the CEM3360 datasheet.
-	constexpr float CEM3360_MAX_GAIN_CV = 1.93F;
-	return std::clamp(cv, 0.0F, CEM3360_MAX_GAIN_CV) / CEM3360_MAX_GAIN_CV;
-}
-
-
 DEFINE_DEVICE_TYPE(VA_VCA,         va_vca_device,         "va_vca",         "Voltage-controlled amplifier")
 DEFINE_DEVICE_TYPE(CA3280_VCA,     ca3280_vca_device,     "ca3280_vca",     "CA3280-based VCA")
 DEFINE_DEVICE_TYPE(CA3280_VCA_LIN, ca3280_vca_lin_device, "ca3280_vca_lin", "Linearized CA3280-based VCA")
-DEFINE_DEVICE_TYPE(CEM3360_VCA,    cem3360_vca_device,    "cem3360_vca",    "CEM3360-based VCA")
