+++ src/devices/sound/va_vca.h
@@ -27,9 +27,6 @@ Summary of devices. See class comments for more details.
 * CA3280_LIN_VCA: CA3280-based VCA in linearized configuration (Id pin connected
   to a current source, or to a resistor to V+). The gain CV is the control
   current supplied to the Iabc pin.
-
-* CEM3360_VCA: A CEM3360 OTA configured as a VCA. The gain CV is the voltage
-  applied to certain pins, depending on configuration (see class documentation).
 */
 
 #ifndef MAME_SOUND_VA_VCA_H
@@ -40,7 +37,6 @@ Summary of devices. See class comments for more details.
 DECLARE_DEVICE_TYPE(VA_VCA, va_vca_device)
 DECLARE_DEVICE_TYPE(CA3280_VCA, ca3280_vca_device)
 DECLARE_DEVICE_TYPE(CA3280_VCA_LIN, ca3280_vca_lin_device)
-DECLARE_DEVICE_TYPE(CEM3360_VCA, cem3360_vca_device)
 
 
 // Emulates a voltage-controlled amplifier (VCA). The control value (CV) can
@@ -207,32 +203,6 @@ private:
 	float m_input_scale;
 	float m_output_scale;
 	float m_cv_scale;
-};
-
-
-// Emulates a CEM3360-based VCA. Each CEM3360 has two OTAs that can be
-// configured as VCAs.
-//
-// The CEM3360 can be configured for linear or exponential gain control.
-// In linear control mode, pin Vo (4, 11) is connected to pin Ve (3, 12), and
-// the CV is applied to pin Vc (5, 10).
-// In exponential mode, pins Vo and Vc do not connect to anything, and the CV
-// is applied to pin Ve, which is also connected to pin Vref (8) via a resistor.
-//
-// Currently, only the linear control mode is implemented. The provided CV is
-// the voltage applied to pin Vc (5, 10).
-//
-// Known inaccuracies:
-// - Linear response. The real device likely has a tanh response, given it is
-//   OTA-based.
-// - Exponential gain control is not implemented.
-class cem3360_vca_device : public va_vca_device
-{
-public:
-	cem3360_vca_device(const machine_config &mconfig, const char *tag, device_t *owner, uint32_t clock = 0) ATTR_COLD;
-
-protected:
-	float cv_to_gain(float cv) const override;
 };
 
 #endif  // MAME_SOUND_VA_VCA_H
